feat(noise): port the fulgora_elevation mix chain, oracle-validated to f32 (#27)
Task 9 of the Fulgora port. Adds `makeFulgoraElevation`, transcribed from
`planet-fulgora-map-gen.lua:206-336` plus `fulgora_dunes`, `fulgora_rock` and
`fulgora_scrap_medium`, validated against a real 2.1.14 Fulgora surface at the
same 101 positions as the shared and cells fixtures.

Two findings, both measured; see `docs/noise/fulgora-elevation-NOTES.md`.

**The port fed the noise primitives f64 coordinates.** `fulgora_basis_oil`
first came back 300x above the port's `basisNoise` floor. The noise machine
passes f32 values between expressions, so the coordinate the game samples at is
an f32; narrowing at the primitive's entry fixes it:

| field | f64 coord | f32 coord |
| --- | --- | --- |
| `fulgora_basis_oil` | 2.37e-4 | 7.15e-7 |
| `fulgora_basis` | 6.94e-6 | 2.09e-7 |
| `fulgora_pyramids` | 7.11e-6 | 1.19e-7 |
| `fulgora_spots` | 7.54e-6 | 1.19e-7 |

Every caller before Fulgora passes a RAW world coordinate, which is already
exactly representable in f32, so this is a measured no-op for them -
`oracle-multioctave.seed123456.json` reports the identical 7.153e-7 worst with
and without it. Fulgora is the first caller to pass a DERIVED coordinate. The
narrowing lives in `sumOctaves` and `makeVoronoi`'s `toGrid` rather than at the
Fulgora call sites, on the same reasoning that put `Math.trunc` inside
`makeVoronoi`. Task 8's `pyramids`/`spots` bounds tighten 1e-5 -> 2e-7 as a
result.

**`slider_rescale` is exact math at per-operation f32.** The default islands
size slider is 1, where `slider_rescale(1, 2)` is exactly 1, so none of the 101
positions can test it - the fixture carries a probe at literal slider values
instead. f32 per-op matches 7/7; f64-then-round misses 2; the noise machine's
fastapprox `pow` misses 6. Only s = 0.5 and s = 5 discriminate.

Six planted errors were confirmed to fail the right tests, including a
`sliderRescale` that returns 1 unconditionally - it passes all 101 positions and
fails only the probe block, which is the whole argument for capturing it.

+++ src/noise/voronoiNoise.ts
@@ -633,8 +633,38 @@ export function makeVoronoi(p: VoronoiParams): Voronoi {
    */
   const divisor = gridSize;
 
-  /** The sample position in grid units, where the cell lattice has unit spacing. */
-  const toGrid = (x: number, y: number): [number, number] => [f32(x / divisor), f32(y / divisor)];
+  /**
+   * The sample position in grid units, where the cell lattice has unit spacing.
+   *
+   * **The incoming coordinates are narrowed to f32 first**, for the same reason
+   * `multioctaveNoise`'s `sumOctaves` does it: the noise machine passes f32
+   * values between expressions, so whatever computed `(x, y)` handed this call
+   * an f32, and narrowing here is what makes the port sample the same point the
+   * game did. This is the single entry point for all four ops, so one narrowing
+   * covers `cellId`, `spotNoise`, `facetNoise` and `pyramidNoise`.
+   *
+   * It is a no-op for a raw world coordinate - an integer or quarter tile below
+   * 2^24 is already exactly representable - which is why every voronoi fixture
+   * committed before Fulgora is unaffected. Fulgora is the first caller to pass
+   * a DERIVED coordinate (`fulgora_wx`, and `ox + wobble_x/2` for the spots),
+   * and there it matters:
+   *
+   * | field | f64 coordinate | f32 coordinate |
+   * | --- | --- | --- |
+   * | `fulgora_pyramids` | 7.11e-6 | **1.19e-7** |
+   * | `fulgora_spots` | 7.54e-6 | **1.19e-7** |
+   * | `fulgora_cells` | exact | exact |
+   *
+   * Both continuous ops land on a single f32 ulp. `cell_id` was exact either
+   * way, and that is the expected asymmetry rather than a reason to doubt the
+   * change: it is a discrete argmin, so a sub-ulp coordinate shift almost never
+   * changes WHICH point is nearest - the same reason the absolute-vs-rebased
+   * delta above needed an exact-value test to catch.
+   */
+  const toGrid = (x: number, y: number): [number, number] => [
+    f32(f32(x) / divisor),
+    f32(f32(y) / divisor),
+  ];
 
   /**
    * The sample-to-point delta in grid units, **rebased on the sample's own
